Aggregated Benchmark Result Object

An R6 class for aggregated benchmark results.

Details

An object of this class is not automatically generated by mlr3::BenchmarkResult, instead it should either be constructed from the results of [mlr3::BenchmarkResult]$aggregate, or custom data can be entered as long as this at least includes the column names learner_id (for models) and task_id (for datasets). Direct coercions from mlr3::BenchmarkResult are available, which essentially wrap $aggregate.

References

Janez Demsar, Statistical Comparisons of Classifiers over Multiple Data Sets, JMLR, 2006

Active bindings

data

([data.table::data.table])
Aggregated data.

learners

(character())
Unique learner names.

tasks

(character())
Unique task names.

measures

(character())
Unique measure names.

nlrns

(integers())
Number of learners.

ntasks

(integers())
Number of tasks.

nmeas

(integers())
Number of measures.

nrow

(integers())
Number of rows.

Methods

Public methods

BenchmarkAggr$new()

BenchmarkAggr$print()

BenchmarkAggr$rank_data()

BenchmarkAggr$friedman_test()

BenchmarkAggr$friedman_posthoc()

BenchmarkAggr$crit_differences()

BenchmarkAggr$clone()

Method new()

Creates a new instance of this R6 class.

Usage

BenchmarkAggr$new(dt)

Arguments

dt

(matrix(1))
matrix like object coercable to data.table::data.table, should include column names "task_id" and "learner_id", coerced to factors internally, and at least one measure (numeric).

Method print()

Prints the internal data via data.table::print.data.table.

Usage

BenchmarkAggr$print(...)

Arguments

...

Passed to data.table::print.data.table.

Method rank_data()

Ranks the aggregated data given some measure.

Usage

BenchmarkAggr$rank_data(meas, minimize = TRUE, task = NULL)

Arguments

meas

(character(1))
Measure to rank the data against, should be in $measures. Can be NULL if only one measure in data.

minimize

(logical(1))
Should the measure be minimized? Default is TRUE.

task

(character(1))
If NULL then returns a matrix of ranks where columns are tasks and rows are learners, otherwise returns a one-column matrix of a specified task, should be in $tasks.

Method friedman_test()

Computes Friedman test over all tasks, assumes datasets are independent.

Usage

BenchmarkAggr$friedman_test(meas = NULL, p.adjust.method = NULL)

Arguments

meas

(character(1))
Measure to rank the data against, should be in $measures. If no measure is provided then returns a matrix of tests for all measures.

p.adjust.method

(character(1))
Passed to p.adjust if meas = NULL for multiple testing correction. If NULL then no correction applied.

Method friedman_posthoc()

Posthoc Friedman Nemenyi tests. Computed with PMCMR::posthoc.friedman.nemenyi.test. If global $friedman_test is non-significant then this is returned and no post-hocs computed. Also returns critical difference

Usage

BenchmarkAggr$friedman_posthoc(meas = NULL, p.value = 0.05)

Arguments

meas

(character(1))
Measure to rank the data against, should be in $measures. Can be NULL if only one measure in data.

p.value

(numeric(1))
p.value for which the global test will be considered significant.

Method crit_differences()

Generate critical differences for a critical difference diagram. Primarily should be used internally by autoplot.BenchmarkAggr. FIXME - CONSIDER MOVING TO PRIVATE HELPER FUNCTION - PROBABLY NOT USEFUL IN ISOLATION

Usage

BenchmarkAggr$crit_differences(
  meas = NULL,
  minimize = TRUE,
  p.value = 0.05,
  baseline = NULL,
  test = c("bd", "nemenyi")
)

Arguments

meas

(character(1))
Measure to rank the data against, should be in $measures. Can be NULL if only one measure in data.

minimize

(logical(1))
Should the measure be minimized? Default is TRUE.

p.value

(numeric(1))
p.value for which the global test will be considered significant.

baseline

(character(1))
For test = "bd" a baseline learner to compare the other learners to, should be in $learners.

test

(character(1)))
Should critical differences be computed from Bonferroni-Dunn (bd) or Nemenyi (nemenyi) tests? Default is Bonferroni-Dunn.

Details

Copied from mlr: Bonferroni-Dunn usually yields higher power than Nemenyi as it does not compare all algorithms to each other, but all algorithms to a baseline instead. Critical differences calculated as: $$CD = q_{\alpha} \sqrt{\left(\frac{k(k+1)}{6N}\right)}$$
# nolint Where \(q_\alpha\) is based on the studentized range statistic. See references for details.

Method clone()

The objects of this class are cloneable with this method.

Usage

BenchmarkAggr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
task = tsks(c("boston_housing", "mtcars"))
learns = lrns(c("regr.featureless", "regr.rpart"))
bm = benchmark(benchmark_grid(task, learns, rsmp("cv", folds = 2)))

# direct coercion
# default measure
as.BenchmarkAggr(bm)

#> <BenchmarkAggr> of 4 rows with 2 tasks, 2 learners and 1 measure
#>           task_id       learner_id  regr.mse
#> 1: boston_housing regr.featureless 87.103452
#> 2: boston_housing       regr.rpart  3.581341
#> 3:         mtcars regr.featureless 40.783770
#> 4:         mtcars       regr.rpart 40.783770

# custom measure
as.BenchmarkAggr(bm, msr("regr.rmse"))

#> <BenchmarkAggr> of 4 rows with 2 tasks, 2 learners and 1 measure
#>           task_id       learner_id regr.rmse
#> 1: boston_housing regr.featureless  9.312375
#> 2: boston_housing       regr.rpart  1.883968
#> 3:         mtcars regr.featureless  6.317312
#> 4:         mtcars       regr.rpart  6.317312

# construct manually
BenchmarkAggr$new(bm$aggregate())

#> <BenchmarkAggr> of 4 rows with 2 tasks, 2 learners and 1 measure
#>           task_id       learner_id  regr.mse
#> 1: boston_housing regr.featureless 87.103452
#> 2: boston_housing       regr.rpart  3.581341
#> 3:         mtcars regr.featureless 40.783770
#> 4:         mtcars       regr.rpart 40.783770

# construct from non-mlr object
df = data.frame(task_id = rep(c("A", "B"), each = 5),
                learner_id = paste0("L", 1:5),
                RMSE = runif(10))
BenchmarkAggr$new(df)

#> <BenchmarkAggr> of 10 rows with 2 tasks, 5 learners and 1 measure
#>     task_id learner_id      RMSE
#>  1:       A         L1 0.6271579
#>  2:       A         L2 0.1684146
#>  3:       A         L3 0.4447896
#>  4:       A         L4 0.1646752
#>  5:       A         L5 0.1182768
#>  6:       B         L1 0.3674644
#>  7:       B         L2 0.9465589
#>  8:       B         L3 0.8754108
#>  9:       B         L4 0.2600589
#> 10:       B         L5 0.6572984